+++ src/renderer/components/ReactFlowBox.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-shadow */
 import { Box, useColorModeValue } from '@chakra-ui/react';
 import log from 'electron-log';
-import { DragEvent, memo, useCallback, useEffect, useMemo, useState } from 'react';
+import { DragEvent, memo, useCallback, useEffect, useMemo } from 'react';
 import ReactFlow, {
     Background,
     BackgroundVariant,
@@ -10,7 +10,6 @@ import ReactFlow, {
     EdgeTypes,
     Node,
     NodeTypes,
-    OnConnectStartParams,
     OnEdgesChange,
     OnNodesChange,
     Viewport,
@@ -19,15 +18,13 @@ import ReactFlow, {
     useReactFlow,
 } from 'react-flow-renderer';
 import { useContext, useContextSelector } from 'use-context-selector';
-import { EdgeData, NodeData, NodeSchema } from '../../common/common-types';
-import { createUniqueId, parseHandle } from '../../common/util';
+import { EdgeData, NodeData } from '../../common/common-types';
 import { AlertBoxContext, AlertType } from '../contexts/AlertBoxContext';
 import { ContextMenuContext } from '../contexts/ContextMenuContext';
-import { GlobalContext, GlobalVolatileContext, NodeProto } from '../contexts/GlobalNodeState';
+import { GlobalContext, GlobalVolatileContext } from '../contexts/GlobalNodeState';
 import { SettingsContext } from '../contexts/SettingsContext';
 import { DataTransferProcessorOptions, dataTransferProcessors } from '../helpers/dataTransfer';
 import { expandSelection, isSnappedToGrid, snapToGrid } from '../helpers/reactFlowUtil';
-import { UseContextMenu } from '../hooks/useContextMenu';
 import { usePaneNodeSearchMenu } from '../hooks/usePaneNodeSearchMenu';
 
 const compareById = (a: Edge | Node, b: Edge | Node) => a.id.localeCompare(b.id);
@@ -322,124 +319,7 @@ const ReactFlowBox = memo(({ wrapperRef, nodeTypes, edgeTypes }: ReactFlowBoxPro
         [createNode, wrapperRef.current, reactFlowInstance]
     );
 
-    const [connectingFrom, setConnectingFrom] = useState<OnConnectStartParams | null>();
-    const [connectingFromType, setConnectingFromType] = useState<string | null>();
-    const [isStoppedOnPane, setIsStoppedOnPane] = useState<boolean>(false);
-
-    useEffect(() => {
-        if (connectingFrom) {
-            const { nodeId, inOutId } = parseHandle(connectingFrom.handleId!);
-            const node = nodes.find((n) => n.id === nodeId);
-            if (node) {
-                const nodeSchema = schemata.get(node.data.schemaId);
-                if (connectingFrom.handleType === 'source') {
-                    const outputType = nodeSchema.outputs[inOutId]?.type;
-                    setConnectingFromType(outputType);
-                } else if (connectingFrom.handleType === 'target') {
-                    const inputType = nodeSchema.inputs[inOutId]?.type;
-                    setConnectingFromType(inputType);
-                } else {
-                    log.error(`Unknown handle type: ${connectingFrom.handleType!}`);
-                }
-            }
-        }
-    }, [connectingFrom]);
-
-    const onPaneContextMenuNodeClick = useCallback(
-        (node: NodeSchema, position: { x: number; y: number }) => {
-            const reactFlowBounds = wrapperRef.current!.getBoundingClientRect();
-            const { x, y } = position;
-            const projPosition = reactFlowInstance.project({
-                x: x - reactFlowBounds.left,
-                y: y - reactFlowBounds.top,
-            });
-            const nodeId = createUniqueId();
-            const nodeToMake: NodeProto = {
-                id: nodeId,
-                position: projPosition,
-                data: {
-                    schemaId: node.schemaId,
-                },
-                nodeType: node.nodeType,
-            };
-            createNode(nodeToMake);
-            if (isStoppedOnPane && connectingFrom) {
-                if (connectingFrom.handleType === 'source') {
-                    const firstValidHandle = schemata
-                        .get(node.schemaId)!
-                        .inputs.find(
-                            (input) => input.type === connectingFromType && input.hasHandle
-                        )!.id;
-                    createConnection({
-                        source: connectingFrom.nodeId,
-                        sourceHandle: connectingFrom.handleId,
-                        target: nodeId,
-                        targetHandle: `${nodeId}-${firstValidHandle}`,
-                    });
-                } else if (connectingFrom.handleType === 'target') {
-                    const firstValidHandle = schemata
-                        .get(node.schemaId)!
-                        .outputs.find((output) => output.type === connectingFromType)!.id;
-                    createConnection({
-                        source: nodeId,
-                        sourceHandle: `${nodeId}-${firstValidHandle}`,
-                        target: connectingFrom.nodeId,
-                        targetHandle: connectingFrom.handleId,
-                    });
-                } else {
-                    log.error(`Unknown handle type: ${connectingFrom.handleType!}`);
-                }
-            }
-
-            setConnectingFrom(null);
-            closeContextMenu();
-        },
-        [
-            connectingFrom,
-            createConnection,
-            createNode,
-            schemata,
-            connectingFromType,
-            isStoppedOnPane,
-        ]
-    );
-
-    const menu: UseContextMenu = usePaneNodeSearchMenu(
-        connectingFrom!,
-        connectingFromType!,
-        onPaneContextMenuNodeClick
-    );
-
-    const onConnectStart = useCallback(
-        (event: React.MouseEvent, handle: OnConnectStartParams) => {
-            setIsStoppedOnPane(false);
-            setConnectingFrom(handle);
-        },
-        [setConnectingFrom, setIsStoppedOnPane]
-    );
-
-    const [coordinates, setCoordinates] = useState<{ x: number; y: number }>({ x: 0, y: 0 });
-
-    const onConnectStop = useCallback(
-        (event: MouseEvent) => {
-            setIsStoppedOnPane(
-                (event.ctrlKey || event.altKey) &&
-                    String((event.target as Element).className).includes('pane')
-            );
-            setCoordinates({
-                x: event.pageX,
-                y: event.pageY,
-            });
-        },
-        [setCoordinates, setIsStoppedOnPane]
-    );
-
-    useEffect(() => {
-        if (isStoppedOnPane && connectingFrom) {
-            const { x, y } = coordinates;
-            menu.manuallyOpenContextMenu(x, y);
-        }
-    }, [isStoppedOnPane, coordinates, connectingFrom]);
+    const { onConnectStart, onConnectStop, onPaneContextMenu } = usePaneNodeSearchMenu(wrapperRef);
 
     return (
         <Box
@@ -481,10 +361,7 @@ const ReactFlowBox = memo(({ wrapperRef, nodeTypes, edgeTypes }: ReactFlowBoxPro
                 onNodesChange={onNodesChange}
                 onNodesDelete={onNodesDelete}
                 onPaneClick={closeContextMenu}
-                onPaneContextMenu={(event) => {
-                    setConnectingFrom(null);
-                    menu.onContextMenu(event);
-                }}
+                onPaneContextMenu={onPaneContextMenu}
                 onSelectionDragStop={onSelectionDragStop}
             >
                 <Background
